Jwt 인증 필터 작성

+++ src/main/java/com/example/quick_gg/repository/StudentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StudentRepository extends JpaRepository<StudentEntity, Long> {
+public interface StudentRepository extends JpaRepository<StudentEntity, Integer> {
         // 조회 결과가 없을 수 있기 때문에 Optional 사용
         Optional<StudentEntity> findByStudentNumber(String studentNumber);
 
@@ -20,6 +20,6 @@ public interface StudentRepository extends JpaRepository<StudentEntity, Long> {
          boolean existsByStudentNumber(String studentNumber);
 
          // id가 이미 등록되어 있는지 확인
-         boolean existsById(Long id);
+         boolean existsById(Integer id);
 
 }
